requirements.txt
# This is a pip requirement file.  See http://pypi.python.org/pypi/pip
# Utility libraries for development
pep8

# scraping libraries
requests
pyquery

# parsing/transforming/services libraries
python-dateutil==1.5
-e git://github.com/groovecoder/geopy.git#egg=geopy

# data storage libraries
couchdb
